refine lexicon update script.

+++ scripts/bin/query_intent/lexicon_job.rb
@@ -6,71 +6,32 @@ require 'json'
 require 'yaml'
 require 'set'
 
+require_relative "job"
 
-module RestClient
-
-  class << self
-    attr_accessor :timeout
-    attr_accessor :open_timeout
-  end
-
-  def self.post(url, payload, headers={}, &block)
-    Request.execute(:method => :post,
-                    :url => url,
-                    :payload => payload,
-                    :headers => headers,
-                    :timeout=>@timeout,
-                    :open_timeout=>@open_timeout,
-                    &block)
-  end
-
-end
-
-class LexiconJob
-  REST_TIMEOUT = 10
-  REST_OPEN_TIMEOUT = 10
-
-  def initialize(config, property)
-      @config = YAML::load_file CONFIG_FILE
-      host = @config["sf1"]
-      ip = host["ip"]
-      port = host["port"]
-      @uri_prefix = "http://#{ip}:#{port}/sf1r"
-      
-      @property =property
-      directory =@config["directory"]
-      directory += "/"
-      @file_path = directory + property
-      
-      @collections = @config["collection"].to_set
-      @hash = Hash.new
-      @toAppend = Array.new
-  end
-
+class LexiconJob < Job
+  
   def run
     read_file()
-    @collections.each do |collection|
+    @@collections.each do |collection|
       request = source_request_template
       request[:collection] = collection
-      #request[:group][:property] = @property
-      response = send_request(request)
+      request[:group].each do |item|
+        item[:property] = @property
+      end
+      response = Job.send_request(request)
       get_labels(response)
     end
     update_file()
-    @collections.each do |collection|
-      request = query_intent_request_template
-      request[:collection] = collection
-      response = send_request(request)
-      if response["header"] && response["header"]["success"]
-        $stdout.print "Reload lexicon successfully for collection: #{collection}\n"
-      end
-    end
   end
 
   protected
 
   def read_file
+    @hash = Hash.new
     @hash.clear()
+    if false == File.exist?(@file_path)
+      return
+    end
     file = File.open(@file_path, "r")
     while(line = file.gets)
       line.chomp!()
@@ -88,25 +49,8 @@ class LexiconJob
     @toAppend.clear()
   end
 
-  def send_request(request)
-    uri = get_uri(request)
-    RestClient.open_timeout = REST_OPEN_TIMEOUT
-    RestClient.timeout = REST_TIMEOUT
-    result = RestClient.post(uri, request.to_json, :content_type => :json)
-    @response = JSON.parse(result)    
-  end
-
-  def get_uri(request)
-    header = request[:header]
-    return nil if header.nil?
-
-    collection = request[:collection]
-    controller = header[:controller]
-    action = header[:action]
-    @uri = "#{@uri_prefix}/#{controller}/#{action}"
-  end
-
   def get_labels(response)
+    @toAppend = Array.new
     group = response["group"];
     group.each do |item|
         labels = item["labels"]
@@ -117,15 +61,6 @@ class LexiconJob
             end
         end
     end
-  end
-  def query_intent_request_template
-    @request = {
-      :header=> {
-        :controller=> "query_intent",
-        :action=> "reload"
-      },
-      :collection=> ""
-   }
   end
 
   def source_request_template
